Add option to build from local ZIP file
The intent is to allow a developer to modify build.txt locally, after
the initial download, in order to shorten the test / change / debug 
cycle.

Additionally, it would allow a developer to download the ZIP file to her
PC, transfer it to a machine without internet access, and build it on
that machine.